move block signature creation related types/function in own module
