+++ cardano/src/block/normal.rs
@@ -6,6 +6,7 @@ use std::collections::{BTreeMap, btree_map};
 use cbor_event::{self, de::RawCbor};
 use super::types;
 use super::types::{HeaderHash, HeaderExtraData, EpochSlotId, ChainDifficulty, SscProof};
+use super::sign::BlockSignature;
 use super::update;
 
 #[derive(Debug, Clone)]
@@ -618,122 +619,6 @@ impl cbor_event::de::Deserialize for Block {
         let body   = raw.deserialize()?;
         let extra  = raw.deserialize()?;
         Ok(Block::new(header, body, extra))
-    }
-}
-
-type SignData = ();
-
-type ProxyCert = hdwallet::Signature<()>;
-
-#[derive(Debug, Clone)]
-pub struct ProxySecretKey {
-    pub omega: u64,
-    pub issuer_pk: hdwallet::XPub,
-    pub delegate_pk: hdwallet::XPub,
-    pub cert: ProxyCert,
-}
-
-impl cbor_event::se::Serialize for ProxySecretKey {
-    fn serialize<W: ::std::io::Write>(&self, serializer: cbor_event::se::Serializer<W>) -> cbor_event::Result<cbor_event::se::Serializer<W>> {
-        serializer.write_array(cbor_event::Len::Len(4))?
-            .serialize(&self.omega)?
-            .serialize(&self.issuer_pk)?
-            .serialize(&self.delegate_pk)?
-            .serialize(&self.cert)
-    }
-}
-
-impl cbor_event::de::Deserialize for ProxySecretKey {
-    fn deserialize<'a>(raw: &mut RawCbor<'a>) -> cbor_event::Result<Self> {
-        raw.tuple(4, "ProxySecretKey")?;
-
-        let omega = cbor_event::de::Deserialize::deserialize(raw)?;
-        let issuer_pk = cbor_event::de::Deserialize::deserialize(raw)?;
-        let delegate_pk = cbor_event::de::Deserialize::deserialize(raw)?;
-        let cert = cbor_event::de::Deserialize::deserialize(raw)?;
-
-        Ok(ProxySecretKey { omega, issuer_pk, delegate_pk, cert })
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct ProxySignature {
-    pub psk: ProxySecretKey,
-    pub sig: hdwallet::Signature<()>,
-}
-
-impl cbor_event::se::Serialize for ProxySignature {
-    fn serialize<W: ::std::io::Write>(&self, serializer: cbor_event::se::Serializer<W>) -> cbor_event::Result<cbor_event::se::Serializer<W>> {
-        serializer.write_array(cbor_event::Len::Len(2))?
-            .serialize(&self.psk)?
-            .serialize(&self.sig)
-    }
-}
-
-impl cbor_event::de::Deserialize for ProxySignature {
-    fn deserialize<'a>(raw: &mut RawCbor<'a>) -> cbor_event::Result<Self> {
-        raw.tuple(2, "ProxySignature")?;
-
-        let psk = cbor_event::de::Deserialize::deserialize(raw)?;
-        let sig = cbor_event::de::Deserialize::deserialize(raw)?;
-
-        Ok(ProxySignature { psk, sig })
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum BlockSignature {
-    Signature(hdwallet::Signature<SignData>),
-    ProxyLight(Vec<cbor_event::Value>), // TODO: decode
-    ProxyHeavy(ProxySignature),
-}
-impl BlockSignature {
-    pub fn to_bytes<'a>(&'a self) -> Option<&'a [u8;hdwallet::SIGNATURE_SIZE]> {
-        match self {
-            BlockSignature::Signature(s) => Some(s.to_bytes()),
-            _ => None,
-        }
-    }
-}
-impl cbor_event::se::Serialize for BlockSignature {
-    fn serialize<W: ::std::io::Write>(&self, serializer: cbor_event::se::Serializer<W>) -> cbor_event::Result<cbor_event::se::Serializer<W>> {
-        match self {
-            &BlockSignature::Signature(ref sig) => {
-                serializer.write_array(cbor_event::Len::Len(2))?
-                    .write_unsigned_integer(0)?.serialize(sig)
-            },
-            &BlockSignature::ProxyLight(ref v) => {
-                let serializer = serializer.write_array(cbor_event::Len::Len(2))?
-                    .write_unsigned_integer(1)?;
-                cbor_event::se::serialize_fixed_array(v.iter(), serializer)
-            },
-            &BlockSignature::ProxyHeavy(ref v) => {
-                serializer.write_array(cbor_event::Len::Len(2))?
-                    .write_unsigned_integer(2)?
-                    .serialize(v)
-            },
-        }
-    }
-}
-impl cbor_event::de::Deserialize for BlockSignature {
-    fn deserialize<'a>(raw: &mut RawCbor<'a>) -> cbor_event::Result<Self> {
-        raw.tuple(2, "BlockSignature")?;
-        let sum_type_idx = raw.unsigned_integer()?;
-        match sum_type_idx {
-            0 => {
-                Ok(BlockSignature::Signature(raw.deserialize()?))
-            },
-            1 => {
-                Ok(BlockSignature::ProxyLight(raw.deserialize()?))
-            },
-            2 => {
-                Ok(BlockSignature::ProxyHeavy(
-                    cbor_event::de::Deserialize::deserialize(raw)?))
-            },
-            _ => {
-                Err(cbor_event::Error::CustomError(format!("Unsupported BlockSignature: {}", sum_type_idx)))
-            }
-        }
     }
 }
